Many changes!
* Add ability to iterate over multiple delta and compression algorithms
and to mix them within a patch.
* Add a new example tool, CompressionDiviner, to introspect compression
parameters from existing archives
* Add a new example tool, CompressionDivinerCompatibilityCheck, to check
whether or not the current runtime matches the expectations of the
CompressionDiviner. If it does, then the output of CompressionDiviner
can be trusted on the current runtime.
* Various code cleanups

As a result of the ability to mix delta and compression algorithms
within the patch, the patch magic has been changed from ARCPAT01 to
ARCPAT02. An attempt has been made to maintain backwards compatibility
even though there has never been a public release of the tool.