Add ability for CompressionDiviner to output a replay file.
The CompressionDiviner can now output a ".directives" companion file for an
archive that details all of the information that it was able to divine about the
archive. Assuming that the information was successfully divined for all entries,
it is possible to exactly replicate the original archive's contents by using the
data from this file.