Add a Reassembler sample that can rebuild an archive using hints from a Diviner.
Use the Diviner to generate a directives file that contains deflate config data
for each entry an archive. The Reassembler can read an archive and use those
directives to reproduce the original archive (for most normal archives). Stats
are tracked to measure the time spent compressing and copying data. This gives a
reasonable estimate of how much effort is required to build the archive, and
provides raw data for performance analysis. Data is sliced a few ways - by file
suffix and also by compression technique, and data that is NOT compressed is
split into categories explaining why it wasn't compressed.