Tighten up batch processing cleanup so it fails less often.
Previously the script could end up deleting files before the service had
finished with them, resulting in filesystem errors on the device (but no data
corruption). Add code to release the CSV stats file in the service and make the
script attempt a gentle shutdown of the service with some grace time to finish
up the files before it tries to clean up.