Greatly increase the data in output from the Reassembler.
The tool now outputs more information at a finer level of detail, in a
slightly more pleasant-to-read format as well.

The tool also can now take an optional --detailed-csv parameter which,
if set, will provide much of this new detail in a CSV format. The data
isn't the exact same set of tuples as in the human-readable report
because it is intended for a different use, e.g. it doesn't include the
per-suffix breakdowns because that would make merging results from many
runs difficult in most row-based CSV importers.