Add tools to transform archives into a more delta-efficient space.
1. UncompressTransformerTool
This tool uses the output from the CompressionDiviner to create a
version of the input archive where all deterministically-recompressible
content is stored uncompressed. The resulting archive is structurally
valid; the compression settings and offsets are updated throughout the
archive, and the data itself is no longer compressed, but aside from
this the archive is identical to the input (extras, comments, etc). An
accompanying report is generated that details the original offsets and
compression settings of the updated entries along with the SHA256 of the
original and transformed archives, which can be used to invert the
transformation with the CompressTransformTool (see below).

It is important to note that although the resulting archive is
structurally valid in every way, it may no longer be suitable for its
original use cases. For example Android application package files (APKs)
often require resources to be aligned on word or page boundaries. This
property is NOT preserved during the transformation. Never fear, the
inverse transformation will fully restore it.

2. CompressTransformerTool
This tool uses the output from the UncompressTransformerTool to rebuild
an original archive from its transformed state. Original compression
settings and offsets are restored, and the SHA256 of the resulting
archive is verified to match the SHA256 of the original archive that
served as input the the UncompressTransformerTool.

A unified "super tool" and accompanying Gradle build file will be
created shortly.