Initiating cleanup for v1.0.
This is a ground-up restructuring, refactoring and pruning of the codebase that
will take place in a few phases. At an extremely high level, it will be like
this:
1. Compatibility checking tools
2. Patch applier
3. Patch generator

Along the way, code will be pulled in (revived) from master, massaged to fit
into the new structure, and thoroughly tested. When this is all done, this
branch will effectively replace master.

This commit prunes almost all of the project, leaving only the compatibility
checking code and renaming it to DefaultDeflateCompatibilityWindow. Everything
that remains alive in code supports this tool with the sole exception of the
uncompressor, which I am leaving in for parity with the compressor interface.