Add a patch explainer tool and accompanying logic.
The PreDiffPlan has been augmented to include the recommendations that were made
along with the reasoning behind them. The PreDiffPlanner now computers and
stores all of that information. This data in turn serves as the basis for a new
PatchExplainer that estimates the likely size of a patch by computing deltas on
each individual pair of files whose uncompressed bytes have changed. This can be
used to attribute the size of a patch back to the files that are changing in the
new archive.