Switch to DivSuffixSort and fix patchgen bug.
Credit to admo@google.com for investigating DivsuffixSort and making it
work well with the existing framework.

This speeds up patch generation by up to 50%. It also fixes issue 7,
which could cause bad patches to be generated when multiple entries in
the "new" archive were matches for an entry in the "old" archive (e.g.,
a file copied to two paths in the "new" archive).