Add a RecommendationModifier to the patch generation process.
A RecommendationModifier can be used to alter the patch generation plan
prior to building a patch. This allows arbitrary modification of the
plan to suit specific needs; for example, a TotalRecompressionLimiter is
provided that will limit the total number of bytes that the patch
applier will need to recompress when applying the patch.

Use of a RecommendationModifier is completely optional, and a Builder
pattern has been used to make this clean.